requirements.txt
# Use local for development only, not for deployment.
homeassistant
pyserial-asyncio
voluptuous
